FileDesc adapted and tests added
FileDesc has tests displaying basic usage, and that FileDesc will throw
exceptions if errors occur.

FileDesc methods will probably not be used in the project for
the most part, FileDesc exists mainly just as a component of the future Socket class,
which will contain a FileDesc as it's only variable.

The primary advantage of the FileDesc in the context of this project is
that it's destructor will call close on its file descriptor.

Utils::runtime_error was also adjusted to contain a std::string as its
member variable. This just makes it simpler to use, as error messages
can be constructed from strings, and will still be valid when passed to
the runtime_error class (as opposed to using str.c_str() as a parameter,
where the string is dropped during the throw). Usage is seen in FileDesc
exceptions (for write and read errors).