Added Socket Module
Socket Module is incomplete, but the idea is for a struct to wrap a
FileDesc.

An IpAddr interface class (with only one current concrete class,
Ipv4Addr) will be used by Socket during instantiation.

This Socket class will eventually be used inside a TcpListener class,
which will hopefully form the backbone of communications in WebServ.