Socket added. Changes to FileDesc and SocketAddrV4
Changes:
* FileDesc copy-constructor edited to simulate move-semantics. Required
the use of const_cast, which I believe to be justified in this case.
Prevents file descriptors being closed upon copy.
* Tests added to FileDesc to demonstrate move semantics behaviour.
* Utils::rvalue class added - I was considering using this for the move
semantics of FileDesc, and it's still possible to do so, but I chose the
const_cast route as it allowed for more implicit code. Whether this is
better or not is something I'll have to discuss with Stan. No tests are
present for this class yet.
* Socket class added. The Socket class uses a SocketAddr class for its
construction, but can also take a const char*, which can be used to
construct a SocketAddrV4 via implicit-conversion to a Str.
* The Socket class contains only one member variable, a file descriptor.
This value is created from the return value of a call to the socket()
function.
* The creation of a Socket class should be controlled by preprocessor
directives for linux vs MacOS. I'm unsure if tests will pass on Mac,
we'll find out.