Threaded Socket test added
I don't know how reliable this test will be, maybe it doesn't work in
all circumstances, but a test has been added that uses two Socket
instances to communicate a message between threads.

The test highlights the need to wrap Socket's into something like a
TcpListener, as the code to test is currently extremely untidy.