Begun work on TcpListener
Also working on TcpStream, which is very similar, but should be used as
an interface to the client file descriptor, rather than as an interface
to the socket file descriptor.

Code isn't prepared yet, my current goal was:

* create a new constructor for SocketAddrV4 that uses a sock_storage as
it's parameter.
* Then, create the accept function in TcpListener, which will return a
TcpStream, which can then be used to interface with a client.