Basic TcpListener implemented and tested.
TcpListener and TcpStream classes have been created. Each contains a
Socket as an inner member variable.

TcpListener has a named constructor 'bind' which will take a SocketAddr
(or a string that can be coerced into one), and then bind to the
relevant port and start listening for connections.

To call accept on the the socket, the accept() member function for
TcpListener has been implemented. It will return a TcpStream object
referring to the connection with the client.

TcpStream has a named constructor 'connect', which calls connect on the
SocketAddr supplied as a parameter.

TcpStream's, once established, can be written to and read from.

Additional changes:
* 'count()' member function added to Str class. Utilised in SocketAddr
construction, which will now reject strings containing more than one
':'.
* Tests for TcpListener added, based on the previous tests in Socket.
Code using TcpListener is much cleaner.