Utils::optional class added and tested
Utils::optional also contains a README.md describing its use in
src/Option.  It is extremely inefficient as compared to std::optional due to
the lack of move semantics in C++98, but it will serve just fine for this
project.

Tests are present, and display the use of optional well.

Plans:
* Use similar implementation as in optional for a Utils::result class,
which can store either the value of a successful operation, or an error
value (most typically a string).

Other changes:
* A fix to the Socket.cpp test, calling SO_REUSEADDR with setsockopt to
ensure tests can be run back to back without relying on the kernel to
clean the port between tests.