Result class written, and minor other changes
The 'result' class has been written - based off of the implementation of
optional. result can either contain a value, or an error.

In practice, result will be used to provide a more modern approach to
error handling (as opposed to returning a special error value of the
same return type, or throwing and catching an exception). A function
should return a Result containing either a successful return value, or
an error value (generally a std::string should fit this purpose).

There aren't many tests currently for the result type - but as it will
replace the exceptions currently in use all throughout the other code,
it should end up being tested all over.

Additional changes:
* remove_const was moved from optional to Utils/src/type_traits.hpp, as
it is used in optional and the new result class.