NON-FUNCTIONAL COMMIT (Union errors in optional)
Turns out that 'unrestricted unions', a union with any data types beyond
PODs, are not permitted in C++98.

This is annoying to say the least, and I'm going to have to re-evaluate
optional as well. While I was trying to avoid it, I'm going to have to
revisit the implementation of boost::optional and take a look at
boost::variant, in particular, to work out how storage is implemented.