NON-FUNCTIONAL COMMIT, Traits module added
I'm leaving things non-functional as they are, just because I'm
extremely happy with the idea behind the non-functional code in
FileDesc. All that's needed is to find a new way of creating the result
and optional classes.

The traits from Boost are utilised heavily in boost::optional and
boost::variant. I'm attempting to understand and implement only the
necessary components to recreate the storage of these classes.