Result implemented in Ipv4Addr
The code that uses Ipv4Addr, namely SocketAddr, has not yet been updated
to pass errors upwards to the caller, but Ipv4Addr is now using the
result-based approach, and tests have been made much nicer as a result
(although unwrap() does have to be called often in tests).

It's an easy task that will take a significant amount of time to change
all throw implementations to result, but it should be very satisfying to
complete. I'll aim to get it done tomorrow.