NON-FUNCTIONAL COMMIT - Result implementations
No issues currently, there's simply a lot of code to convert from "throw
exception" to "return Result".

So far, I'm up to TcpStream, which is I think the final class that
requires updating. Then, there will be a lot of changes to make to the
test files. Updating the test files should be one of the most satisfying
parts however, as EXPECT_THROW tests can be changed into a comparison of
Result values.

Plans:
* I think I'll make some changes to the .clang-format file soon, namely:
        * Separate template parameters from concept
        * Reduce column limit from 120 to 100