Result fully implemented
After completing the Utils::result class, it has now been utilised in
all of the existing code written for the Socket module.

All tests are passing, and it's prepared for a pull request - the next
PRs will probably involve fixing bugs to pass all the tests in the
Github actions environment as well.