TcpListener::select complete. Examples dir added
TcpListener now also has the functionality to handle clients
asynchronously using TcpListener::select(). This function wraps around
the select function from the C standard library, to monitor connections
in a single thread and only respond to those connections that have sent
data to the server.

I've also added an Examples directory. The testing of code so far has
been excellent, and should continue (and be expanded upon with Google
Mock) in the future, however I thought it would be sensible to supply
test executables for some particular areas of code, to demonstrate their
usage. Select.run is one such executable, it demonstrates an extremely
simple TCP server that will read input from clients, print it to the
screen, and send back a message.