Unit test added for TcpListener::select
Additionally, some extra functionality was added to TcpStream. Namely,
the ability to read into a std::string by reference from a Socket.
I'm not positive if I'm using the best possible algorithm for
performance for this function, but it does the job.

TcpListener required practically no changes in order to pass the test -
which creates a thread that generates three further threads to send
messages to the specified address.