Select (#28)
* TcpListener::select complete. Examples dir added

TcpListener now also has the functionality to handle clients
asynchronously using TcpListener::select(). This function wraps around
the select function from the C standard library, to monitor connections
in a single thread and only respond to those connections that have sent
data to the server.

I've also added an Examples directory. The testing of code so far has
been excellent, and should continue (and be expanded upon with Google
Mock) in the future, however I thought it would be sensible to supply
test executables for some particular areas of code, to demonstrate their
usage. Select.run is one such executable, it demonstrates an extremely
simple TCP server that will read input from clients, print it to the
screen, and send back a message.

* Bug fix for MacOS

* Unit test added for TcpListener::select

Additionally, some extra functionality was added to TcpStream. Namely,
the ability to read into a std::string by reference from a Socket.
I'm not positive if I'm using the best possible algorithm for
performance for this function, but it does the job.

TcpListener required practically no changes in order to pass the test -
which creates a thread that generates three further threads to send
messages to the specified address.

* MacOS bug fix

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>