Result implement (#26)
* Result implemented in Ipv4Addr

The code that uses Ipv4Addr, namely SocketAddr, has not yet been updated
to pass errors upwards to the caller, but Ipv4Addr is now using the
result-based approach, and tests have been made much nicer as a result
(although unwrap() does have to be called often in tests).

It's an easy task that will take a significant amount of time to change
all throw implementations to result, but it should be very satisfying to
complete. I'll aim to get it done tomorrow.

* NON-FUNCTIONAL COMMIT - Result implementations

No issues currently, there's simply a lot of code to convert from "throw
exception" to "return Result".

So far, I'm up to TcpStream, which is I think the final class that
requires updating. Then, there will be a lot of changes to make to the
test files. Updating the test files should be one of the most satisfying
parts however, as EXPECT_THROW tests can be changed into a comparison of
Result values.

Plans:
* I think I'll make some changes to the .clang-format file soon, namely:
        * Separate template parameters from concept
        * Reduce column limit from 120 to 100

* Result fully implemented

After completing the Utils::result class, it has now been utilised in
all of the existing code written for the Socket module.

All tests are passing, and it's prepared for a pull request - the next
PRs will probably involve fixing bugs to pass all the tests in the
Github actions environment as well.

* Attempted MacOS bug fix

Forgot to update Socket::init code for MacOS

* Formatting changes (#27)

* Changes to Clang Format

* Lines now limited to 100 chars width.
* Template declarations now break onto the second line.

* Select (#28)

* TcpListener::select complete. Examples dir added

TcpListener now also has the functionality to handle clients
asynchronously using TcpListener::select(). This function wraps around
the select function from the C standard library, to monitor connections
in a single thread and only respond to those connections that have sent
data to the server.

I've also added an Examples directory. The testing of code so far has
been excellent, and should continue (and be expanded upon with Google
Mock) in the future, however I thought it would be sensible to supply
test executables for some particular areas of code, to demonstrate their
usage. Select.run is one such executable, it demonstrates an extremely
simple TCP server that will read input from clients, print it to the
screen, and send back a message.

* Bug fix for MacOS

* Unit test added for TcpListener::select

Additionally, some extra functionality was added to TcpStream. Namely,
the ability to read into a std::string by reference from a Socket.
I'm not positive if I'm using the best possible algorithm for
performance for this function, but it does the job.

TcpListener required practically no changes in order to pass the test -
which creates a thread that generates three further threads to send
messages to the specified address.

* MacOS bug fix

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>