Result (#25)
* Result class written, and minor other changes

The 'result' class has been written - based off of the implementation of
optional. result can either contain a value, or an error.

In practice, result will be used to provide a more modern approach to
error handling (as opposed to returning a special error value of the
same return type, or throwing and catching an exception). A function
should return a Result containing either a successful return value, or
an error value (generally a std::string should fit this purpose).

There aren't many tests currently for the result type - but as it will
replace the exceptions currently in use all throughout the other code,
it should end up being tested all over.

Additional changes:
* remove_const was moved from optional to Utils/src/type_traits.hpp, as
it is used in optional and the new result class.

* NON-FUNCTIONAL COMMIT (Union errors in optional)

Turns out that 'unrestricted unions', a union with any data types beyond
PODs, are not permitted in C++98.

This is annoying to say the least, and I'm going to have to re-evaluate
optional as well. While I was trying to avoid it, I'm going to have to
revisit the implementation of boost::optional and take a look at
boost::variant, in particular, to work out how storage is implemented.

* NON-FUNCTIONAL COMMIT, Traits module added

I'm leaving things non-functional as they are, just because I'm
extremely happy with the idea behind the non-functional code in
FileDesc. All that's needed is to find a new way of creating the result
and optional classes.

The traits from Boost are utilised heavily in boost::optional and
boost::variant. I'm attempting to understand and implement only the
necessary components to recreate the storage of these classes.

* NON-FUNCTIONAL COMMIT - New optional implementation

The new implementation using Boost 2004's methodology for optional is
passing all tests, and furthermore, it has been tested for compilation
with C++98, and it passes.

Next, is to perform similar alterations for Result, and after testing, I
should finally be able to make another Pull Request.

* NON-FUNCTIONAL COMMIT - Result updated

Utils::result has now been updated the same way that Utils::optional
was, it both passes its tests, and has been tested to compile
successfully with C++98.

Now after the next target, reimplementing a function in FileDesc to work
using a Utils::result rather than a throw for error handling, then I'll
turn this branch into a pull request.

* Result class complete

Result class is now completed, and in use for the first time in
FileDesc.

There are numerous benefits to using the result class, one being that
tests should no longer need to use the difficult EXPECT_THROW syntax when
testing error cases.

In the main code base, result provides the option of handling errors
sensibly, without polluting the codebase with try-catch blocks.

This warrants a PR, and there will be another PR once all existing code
has been adapted to use 'result'.

* Bug fix

-1 value for size_t generates compiler error on github action
environment in type_traits.cpp test file. Changed to positive value.

* Bug fix

TYPED_TEST_CASE in googletest is deprecated. Was in use in Traits test.
Updated to TYPED_TEST_SUITE

* Result implement (#26)

* Result implemented in Ipv4Addr

The code that uses Ipv4Addr, namely SocketAddr, has not yet been updated
to pass errors upwards to the caller, but Ipv4Addr is now using the
result-based approach, and tests have been made much nicer as a result
(although unwrap() does have to be called often in tests).

It's an easy task that will take a significant amount of time to change
all throw implementations to result, but it should be very satisfying to
complete. I'll aim to get it done tomorrow.

* NON-FUNCTIONAL COMMIT - Result implementations

No issues currently, there's simply a lot of code to convert from "throw
exception" to "return Result".

So far, I'm up to TcpStream, which is I think the final class that
requires updating. Then, there will be a lot of changes to make to the
test files. Updating the test files should be one of the most satisfying
parts however, as EXPECT_THROW tests can be changed into a comparison of
Result values.

Plans:
* I think I'll make some changes to the .clang-format file soon, namely:
        * Separate template parameters from concept
        * Reduce column limit from 120 to 100

* Result fully implemented

After completing the Utils::result class, it has now been utilised in
all of the existing code written for the Socket module.

All tests are passing, and it's prepared for a pull request - the next
PRs will probably involve fixing bugs to pass all the tests in the
Github actions environment as well.

* Attempted MacOS bug fix

Forgot to update Socket::init code for MacOS

* Formatting changes (#27)

* Changes to Clang Format

* Lines now limited to 100 chars width.
* Template declarations now break onto the second line.

* Select (#28)

* TcpListener::select complete. Examples dir added

TcpListener now also has the functionality to handle clients
asynchronously using TcpListener::select(). This function wraps around
the select function from the C standard library, to monitor connections
in a single thread and only respond to those connections that have sent
data to the server.

I've also added an Examples directory. The testing of code so far has
been excellent, and should continue (and be expanded upon with Google
Mock) in the future, however I thought it would be sensible to supply
test executables for some particular areas of code, to demonstrate their
usage. Select.run is one such executable, it demonstrates an extremely
simple TCP server that will read input from clients, print it to the
screen, and send back a message.

* Bug fix for MacOS

* Unit test added for TcpListener::select

Additionally, some extra functionality was added to TcpStream. Namely,
the ability to read into a std::string by reference from a Socket.
I'm not positive if I'm using the best possible algorithm for
performance for this function, but it does the job.

TcpListener required practically no changes in order to pass the test -
which creates a thread that generates three further threads to send
messages to the specified address.

* MacOS bug fix

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>