MacOS bugs fixed
In Socket.cpp: On MacOS, a Socket was being initialized, then a new
socket was being created and returned from the same file descriptor,
causing the original Socket to go out of scope and close it's file
descriptor.

Fixed now.

Additional minor bugs fixed include unused member variable in Client,
and an address that could not be bound to in TcpListener tests