NON-FUNCTIONAL: Beginning HTTP
Made the decision to change the design of the Server - instead of
pre-allocating a large region of memory, and utilising this for all
clients throughout the program, memory allocation will be as in a normal
C++ program - with a single try-catch block placed around the one area
of the code that requires additional memory allocation (reading and
parsing of HTTP requests, & creation of HTTP responses).

This offers far more flexibility when it comes to creating the code for
HTTP Requests, and if an error is ever encountered, the catch-block will
be responsible for preparing a 5xx internal server error response for
the relevant client.

Current development code involves creating a Builder for a Request
class, that will likely be used during the reading and parsing of a
Client's request. If not, it will certainly be valuable for testing
purposes alone.