NON-COMPLETE COMMIT Continuing Request Parsing
Almost complete with the parsing I think, method and headers are now
well parsed - state of parsing is well handled, and different error
messages are created depending on what error is encountered.

Reading of body is relatively simple, as it should normally just simply
append the body that is parsed.

Chunked data is handled slightly differently however. It's almost
handled currently, but requires a slight bit more work.

A couple of changes to Utils were made, primarily the addition of a
atol and atol_length, for parsing number values.