HTTP Parser has basic functionality
The HTTP Request Parser class is now in a basic state of functionality.
Using the Request::Parser, it is possible to create a Parser instance -
using the parse_line function repeatedly, all the logic for HTTP Request
parsing is handled internally (ensuring Header limits are obeyed,
parsing chunked requests, validating presence of CRLF in the start-line
and header fields, etc.).

Additional functionality is likely to be required in future however:
- Functionality to inform the calling code whether the parser
is expecting a Chunked Request or a Message Body, and if so, to use the
parser's stored content-length value to decide how many bytes to read
from the client socket.
- To use a client-body-size limit (defined in config) to limit how many
  characters are parsed by the Parser. (This logic may be better suited
  to the calling code however).