Wrapped HTTP Version and Method enums in classes
As with the Request::Parser::State, wrapped the existing enums into
classes. For the State, it was done so that State could be returned as
an Error value in the Utils::result class - in this instance, it was
done simply to remove the need for silly 'stringify' functions and to
improve the simplicity of the ostream overloads.