FAILING TEST COMMIT - Client Parsing development
The HTTP Parser is effective, now currently development is underway to
implement it in the Client class.

Currently, there's a balancing act of trying to accept input from
multiple terminals at once using netcat (an excellent example of this
working is present in src/Examples/src/ClientParsing.cpp), and to
accept input in one chunk, even when this input doesn't end in a new
line character (as in the failing test, in tst/HttpParsing.cpp).

Next steps:

- Fix failing test
- Create Response class (No Parser required, but a Builder would be
  welcome)
- Add code to check whether a Request is valid beyond the error checking
  to see if it is well formed, e.g. check to see that a requested file
  actually exists.

Optional steps:

- Consider also creating a URI class, to better represent the separate
  components of a URI, instead of just storing it as a single
  std::string.