Begun on Response class - including Builder
A basic builder has been created for a http::Response class. This
functions much the same way as the builder for the http::Request class -
with minimal/virtually no error checking.

The builder has some tests, but as there isn't a Parser for this class,
tests aren't as extensive as for the http::Request.

http::Request::State has been moved out of Request to its own file in
the http namespace, as it is shared between Requests and Responses.

The enum_strings of the State enum class have been changed to fit with
the exact title of the responses as found on
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status.
Tests have been changed as appropriate.

A small class named Time has been added inside of Utils namespace. There
was an intention to give it more functionality, but all it really will
serve to do is to supply the time in the correct Date format with the
function to_http_string.
Test for this is incomplete, but is located in
src/Utils/tst/time_tests.cpp.