Basic LayerConfig functionality
Alternative approach to the configuration data structure.

Utilises nested instances of the same class - Layer, by having the root
layer place its children inside of a vector.

A key benefit of the Layer class is that multiple children can be placed
inside of the same instance without naming conflicts, as children are
identified by a UID, comprising of their name, depth in tree-structure,
and an uint8_t identifier.

Upon placing a new layer into a tree, a Result is returned that upon
success, contains a pointer to the newly formed Layer.

Each layer is also capable of creating a "Layer::Location" struct, which
in future can be used to traverse the tree from any Layer to the
requested Layer.

A basic to_string() function was created for testing - but shouldn't be
used in production.

A Layer::Iterator nested class has also been created, and will be used
as a normal iterator through the children of a layer, but will contain
an identifier that filters out child layers that don't have the
requested name. This class is still in development.