Layer Iterator functional
An iterator over child layers within a layer is now present - it can be
used to iterator over every child layer, or only layers that have a
specified name as identifier.

In our project, this will be primarily useful for iterating over the
'server' blocks to produce the relevant number of TcpListeners during
set up of the server based on the config file.