Layer development complete
Development of the Layer Class is complete for now.
The next step, is to create a Config class (which will contain a Layer
instance, and not much else) and a Config::Parser.

The Config::Parser should lean heavily upon the BufReader and Slice
classes to efficiently read from a file - verify that lines of the file
are valid, and fill the Config struct with the data.