Begun work on parsing Config file
Beginning with parsing, but no major progress yet.

So far, comments are handled - and will not form a part of processing.
Whitespace is successfully ignored as well.

The major issue is: nginx config files are still valid when a "{" is
located directly within a string of characters, and does not require any
whitespace between the directive and the brace.

Additionally, key value pairs may be contained on the same line as a
directive. Effectively - all whitespace is ignored in an nginx
configuration file - and values require separation by ";".

These issues should hopefully be minor, but they are annoyances
nonetheless. Next commit will hopefully contain a working parser.