Config file parsing complete
Configuration files can now be read into a data structure named Layer.

Layer is comprised of a 'global' layer, that represents the global
configuration, and multiple layer blocks within it.

Currently, any directive is accepted as the name for a block, http and
server are examples of valid directives, but the parser currently allows
anything.

If the 'include' directive is specified anywhere, then the value will be
treated as a filename for another config file, and this will be read
appropriately into the data structure. This does not currently handle
wildcards, and currently only accepts one filename per include
directive.

Tests are not yet extensive, and so far only rely on displaying the data
structure to the screen with an ostream overload. More explicit tests
will follow.

Minor changes have been made to Utils::result and Slice.