Http (#37)
* BufReader and File added

Two new classes, File to wrap around a FileDesc for an open file, and
BufReader, which wraps around a Reader to enable efficient read
operations.

* Basic Client and Server created

Client and Server classes have now been created - they wrap a lot of
pre-existing code, and should hopefully form the base of the WebServ
project.

Server now asynchronously handles clients, maintaining connection until
they are closed by the client - however changes will be required in
future to disconnect clients if the limit on clients is reached.

src/Examples/src/Select.cpp has been altered - previously it provided an
example of a Server using a single TcpListener and TcpStream's. Given
the requirements of the project (no memory allocation after program
initialization), this example has been changed to show the new Server
implementation.

* MacOS bugs fixed

In Socket.cpp: On MacOS, a Socket was being initialized, then a new
socket was being created and returned from the same file descriptor,
causing the original Socket to go out of scope and close it's file
descriptor.

Fixed now.

Additional minor bugs fixed include unused member variable in Client,
and an address that could not be bound to in TcpListener tests

* Fixed private-public mistake in BufReader

* BufReader is updated with eof() function

Implementation is very simple for a BufReader<File>, since you can
always read from a file until EOF. For a BufReader<TcpStream> however,
there is no acceptable solution, without some control by the Client
class that is being used to encapsulate the BufReader<TcpStream>.

This is obviously not ideal, but it currently does result in an
extremely tractable API with the client for reading Requests line by
line.

* Removed potentially erroneous line in BufReader

* NON-FUNCTIONAL: Beginning HTTP

Made the decision to change the design of the Server - instead of
pre-allocating a large region of memory, and utilising this for all
clients throughout the program, memory allocation will be as in a normal
C++ program - with a single try-catch block placed around the one area
of the code that requires additional memory allocation (reading and
parsing of HTTP requests, & creation of HTTP responses).

This offers far more flexibility when it comes to creating the code for
HTTP Requests, and if an error is ever encountered, the catch-block will
be responsible for preparing a 5xx internal server error response for
the relevant client.

Current development code involves creating a Builder for a Request
class, that will likely be used during the reading and parsing of a
Client's request. If not, it will certainly be valuable for testing
purposes alone.

* NON-COMPLETE COMMIT: Developing HTTP Request

1. Request::Builder is mostly created. Contains practically no error
   checking, this structure is just responsible for creating a Request.
2. Request::Parser contains a Builder, and much more error handling to
   determine what stage of parsing a request we are at.

   More to follow

* NON-COMPLETE COMMIT - Still working on Request

HTTP Request is progressing excellently - A Request::Builder is wrapped
effectively inside of a Request::Parser for parsing requests line by
line.

The Str class required some bolstering before moving forward with the
Request parsin however, which this commit contains. Changes include:

- Improvements to readability by adding a private function to the Split
  class named is_from_slice() - this is used to check whether the Split
  was created from another Str (slice), or from a raw char* string, as
  the logic should be different for the Split depending on source.
- Now Str::SplitN class added - practically the same as the Str::Split
  class, but with an extra parameter that will limit the number of
  splits possible. This can be used to split "Accept: this: line with
  :colons" into "Accept" and "this: line with :colons", instead of
  splitting on every ":", the choice is available to split into however
  many values are desired. This as an obvious application in parsing
  Http headers.

* NON-COMPLETE COMMIT Continuing Request Parsing

Almost complete with the parsing I think, method and headers are now
well parsed - state of parsing is well handled, and different error
messages are created depending on what error is encountered.

Reading of body is relatively simple, as it should normally just simply
append the body that is parsed.

Chunked data is handled slightly differently however. It's almost
handled currently, but requires a slight bit more work.

A couple of changes to Utils were made, primarily the addition of a
atol and atol_length, for parsing number values.

* HTTP Parser has basic functionality

The HTTP Request Parser class is now in a basic state of functionality.
Using the Request::Parser, it is possible to create a Parser instance -
using the parse_line function repeatedly, all the logic for HTTP Request
parsing is handled internally (ensuring Header limits are obeyed,
parsing chunked requests, validating presence of CRLF in the start-line
and header fields, etc.).

Additional functionality is likely to be required in future however:
- Functionality to inform the calling code whether the parser
is expecting a Chunked Request or a Message Body, and if so, to use the
parser's stored content-length value to decide how many bytes to read
from the client socket.
- To use a client-body-size limit (defined in config) to limit how many
  characters are parsed by the Parser. (This logic may be better suited
  to the calling code however).

* Integration test prepared

The first integration test - using a client in one thread to send a HTTP
Request, and a Server in another thread to parse the request - has been
prepared (skeleton draft).

Next step is to incorporate the HTTP Request Parsing into a client, then
setup the test.

* Wrapped HTTP Version and Method enums in classes

As with the Request::Parser::State, wrapped the existing enums into
classes. For the State, it was done so that State could be returned as
an Error value in the Utils::result class - in this instance, it was
done simply to remove the need for silly 'stringify' functions and to
improve the simplicity of the ostream overloads.

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>
Co-authored-by: Ryan Lucas <rlucas@f1r4s12.codam.nl>