Client request (#41)
* Removed unnecessary data variable in Client

* FAILING TEST COMMIT - Client Parsing development

The HTTP Parser is effective, now currently development is underway to
implement it in the Client class.

Currently, there's a balancing act of trying to accept input from
multiple terminals at once using netcat (an excellent example of this
working is present in src/Examples/src/ClientParsing.cpp), and to
accept input in one chunk, even when this input doesn't end in a new
line character (as in the failing test, in tst/HttpParsing.cpp).

Next steps:

- Fix failing test
- Create Response class (No Parser required, but a Builder would be
  welcome)
- Add code to check whether a Request is valid beyond the error checking
  to see if it is well formed, e.g. check to see that a requested file
  actually exists.

Optional steps:

- Consider also creating a URI class, to better represent the separate
  components of a URI, instead of just storing it as a single
  std::string.

* Fixed failing test

I think this is a significant amount of changes for a PR now.

- Bug fix: The failing tests was relatively simple to fix, I added code
  in Client during the generate_request() call to read everything left
  from the client until content-length in the parser was 0 - this should
  ensure that the entire body of a client's message is always read, even
  when the body doesn't end in a CRLF.
  Any information after a CRLF is ignored.
- Bug fix: While fixing the aforementioned bug, I happened to run the
  tests through valgrind, and was met with an insane amount of error
  messages concerning "Uninitialized values". Long story short, after a
  couple of painful hours, I learned you can't pass a char* by reference
  into a lambda in C++. Errors vanished after passing by value.
- Bug fix: I found valgrind errors in the Socket connection test.
  Discovered it was due to an uninitialized socklen_t. Fixed now by
  initializing to 0.

Next goals remain the same as before. Next PR will be to do with a
Response class, and there should be testing for this too. As there is no
parsing required for this class however, it should be simpler than the
Request class.

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>
Co-authored-by: Ryan Lucas <rlucas@f1r4s12.codam.nl>