Layer config (#43)
* Basic LayerConfig functionality

Alternative approach to the configuration data structure.

Utilises nested instances of the same class - Layer, by having the root
layer place its children inside of a vector.

A key benefit of the Layer class is that multiple children can be placed
inside of the same instance without naming conflicts, as children are
identified by a UID, comprising of their name, depth in tree-structure,
and an uint8_t identifier.

Upon placing a new layer into a tree, a Result is returned that upon
success, contains a pointer to the newly formed Layer.

Each layer is also capable of creating a "Layer::Location" struct, which
in future can be used to traverse the tree from any Layer to the
requested Layer.

A basic to_string() function was created for testing - but shouldn't be
used in production.

A Layer::Iterator nested class has also been created, and will be used
as a normal iterator through the children of a layer, but will contain
an identifier that filters out child layers that don't have the
requested name. This class is still in development.

* Layer Iterator functional

An iterator over child layers within a layer is now present - it can be
used to iterator over every child layer, or only layers that have a
specified name as identifier.

In our project, this will be primarily useful for iterating over the
'server' blocks to produce the relevant number of TcpListeners during
set up of the server based on the config file.

* Layer development complete

Development of the Layer Class is complete for now.
The next step, is to create a Config class (which will contain a Layer
instance, and not much else) and a Config::Parser.

The Config::Parser should lean heavily upon the BufReader and Slice
classes to efficiently read from a file - verify that lines of the file
are valid, and fill the Config struct with the data.

* Begun work on parsing Config file

Beginning with parsing, but no major progress yet.

So far, comments are handled - and will not form a part of processing.
Whitespace is successfully ignored as well.

The major issue is: nginx config files are still valid when a "{" is
located directly within a string of characters, and does not require any
whitespace between the directive and the brace.

Additionally, key value pairs may be contained on the same line as a
directive. Effectively - all whitespace is ignored in an nginx
configuration file - and values require separation by ";".

These issues should hopefully be minor, but they are annoyances
nonetheless. Next commit will hopefully contain a working parser.

* Config file parsing complete

Configuration files can now be read into a data structure named Layer.

Layer is comprised of a 'global' layer, that represents the global
configuration, and multiple layer blocks within it.

Currently, any directive is accepted as the name for a block, http and
server are examples of valid directives, but the parser currently allows
anything.

If the 'include' directive is specified anywhere, then the value will be
treated as a filename for another config file, and this will be read
appropriately into the data structure. This does not currently handle
wildcards, and currently only accepts one filename per include
directive.

Tests are not yet extensive, and so far only rely on displaying the data
structure to the screen with an ostream overload. More explicit tests
will follow.

Minor changes have been made to Utils::result and Slice.

* Fixed MacOS error

/etc/nginx/mime.types is not necessarily going to exist on a filesystem
that the tests are run on, so the absolute path was changed to a
relative path (in the complex parsing test of LayerConfig).

Co-authored-by: Ryan Lucas <rlucas@student.codam.nl>
Co-authored-by: Ryan Lucas <rlucas@f1r4s12.codam.nl>