Change to Slice::Split and additiona Conf validation
A count_remaining() function was added to Slice::Split - which counts
how many values are remaining in a Split, without the need to collect
them into a vector.

Additional validations for config file parsing are being prepared:
Namely, a std::map that contains the available names of valid
directives, as well as additional information on how many arguments they
should have, and whether they can exist as a Block or as a Key (e.g.,
"http" is an invalid key name, and should only be a block. server_name
on the other hand, cannot represent a block, but should exist only as a
key).

Validations will be implemented to the point of:
- checking if the directive is correctly made a block or a scope.
- checking that the directive has a valid number of values.

Validations could in future be extended to include:
- checking that a directive only exists within a scope that is able to
  exist in (e.g. a "server" block can only exist inside of an "http"
  block).
- checking that a singleton directive does not exist multiple times
  (e.g. you should not have more than one "http" block in an nginx
  configuration (I think)).