EXPERIMENTAL COMMIT
Still thinking of ideas as to how to implement the "server" blocks of
nginx.

My current idea is to go with a "VirtualServer" class, that is
essentially a wrapper around a Layer pointer to a "server" block config.

Then, member functions of the VirtualServer class should offer easy
access to information of the "server" blocks config.

Additionally, the Client class is going to have to store a
server_address variable, which will contain a SocketAddrV4 corresponding
to the TcpListener that it was accepted by. This value can be obtained
using the new TcpListener::get_sock_name() member function.
This value will then be used (once the Request is completed), to
determine which VirtualServer config will be used to process the
Request, by comparing:
 - The SocketAddrV4 values generated by the VirtualServer blocks
 - The server_name of the VirtualServer block, and the Host name of the
   Request
 - The default_server directive of the VirtualServer block.